Register Feature:
Added a method in the Roles Controller to
retreive all roles except the Admin Role

Users Controller:
Created 2 Methods for register one Get and a Post request.

Within the post of Register there's the following validations:
Checks whether email already exists in the database
Hashes inputted password with it's own HashPassword method.

Removed Create methods and cshtml as now we'll use the Register.

Updated the htmls for User for proper types of client side validation.

In _Layout added a line to include Register redirect in the navigation bar