Session Variables:
Added Nu Get Session Variables

Edited Users Index method that only users logged in and are of RoleType Admin allowed inside.
else users logged in are redirect to their details page

Added Login functions to call login page

Added RetreiveSessionUser method to call session variable

Edited Roles in models so the collection users are json ignroed and ignoreData member
in order to not have self looping objects.

In Startup added allowing for session in application

Created a new class SessionHelper under new folder View Models,
purpose of this class is to set an object into a json string to session & get session to json

_Layout.cshtml received small upgrade.
If now includs if user not logged in it displays Register / Login
If user is logged in it changes to Profile (Details of user) and Logout function.

Details.cshtml includes if statement to redirect to users index if user logged in is of roletype Admin